refactor: return [u8; 512]

+++ shower4j/src/main/java/com/erayt/shower4j/ShowerCallback.java
@@ -1,5 +1,5 @@
 package com.erayt.shower4j;
 
 public interface ShowerCallback {
-    void callback(long[] data, int size);
+    void callback(byte[] data, int size);
 }
